Created the structure for the parallel version
There is now a houghTransform .h .cu pair that will perform the preprocessing
and the actual transform. Line extraction will also be situated here. The class
HoughParameterSet has been moved into these files as well. All of this is inside a namespace called cudaHough.
The main-function performs the neccessary steps for extracting the lines.