Reformatted code, sobel convolution
Code has been reformatted such that namespaces and class visibility modifiers
create a new level of indentation.

First approach at a parallelized variant of convolution. No shared memory is
used so far. Implementation of convolution with sobel filters in preprocess
is present, but untested. Calculating the gradient strength still needs to
be done.