Implemented Binarization, threshold to be set
Introduced the binarize(..) method and the Kernel binarizeGPU(..) which take a double image and binarize it according to a threshold. preprocess(..) already makes use of this and now returns the binary image. In accordance with that, preprocess(..) now returns a bool pointer instead of a CImg. The threshold can be passed via the terminal. As of now, this is an absolute value, not a divisor of the range represented in the image.