Fixed accessing wrong dimensions of accumulator array
Previously, we used the dimensions of the original image when moving the accumulator array to ram, and when displaying it on screen. The accumulator has its own dimensions, of course. Now, the HoughParameterSet class features a function to compute these dimensions, and they are used where appropriate.

Additionally, binarized image and accumulator array are now stored to disk.