Early work on blurring in paralell version
Implemented a simple version of the gauss filter for blurring images. This requires another parameter sigma (which as of now is hard-coded and can not be passed as a command line argument). The filter dimensions are automatically derived from sigma. They are 2*sigma in both directions, which stems from the two-sigma rule that 95% of probability mass lies within two sigma of the origin.

Since the gauss function integrates to one, there is no requirement for normalization. However, we could consider adding a constant factor to account for the fact that we lose ~5% of the total sum.